// src/app/enums/radiation-field-enum.ts
export enum RadiationFieldEnum {
    ChestWall = "Chest Wall",
    WholeBreast = "Whole Breast",
    PartialBreast = "Partial Breast",
    Axilla = "Axilla",
    Superclav = "Superclav",
    InternalMammary = "Internal Mammary",
    Spine2Field = "Spine (2-field)",
    SpinePosterior = "Spine (Posterior)",
    WholeBrain = "Whole Brain",
    Other = "Other"
}